move metadata into a separate shard
Previously labels, timestamp, etc. were all in the same shard with the
actual contents. The thinking was that so you can restore a working
shard server (that has lost its dynamic state somehow) from the shards
alone. But this is not a great idea, because you might want to share
your content without sharing the metadata (other people don't
necessarily need to know which labels you used or when it was created).
Separating them also allows to update the metadata (with pointers to the
main shard / previous metadata version in the new one).